% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{convert_integer}
\alias{convert_integer}
\title{Convert elements in table into integer}
\usage{
convert_integer(table_shots)
}
\arguments{
\item{table_shots}{data.table}
}
\value{
data.table
}
\description{
Convert elements in table into integer
}
