% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daten_einlesen.R
\name{delete_character}
\alias{delete_character}
\title{Edit character of Shots}
\usage{
delete_character(x, ...)
}
\arguments{
\item{x}{String. of the form "\"information\"text".}
}
\value{
String. with just the information.
}
\description{
Separate the one column in Shots into multiple columns
and then separate the additional information.
}
