% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_series}
\alias{get_series}
\title{Get series of a Shooter}
\usage{
get_series(shoot_table)
}
\arguments{
\item{shoot_table}{data.table}
}
\value{
data.table
}
\description{
Get series of a Shooter
}
