% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{get_shooting_list}
\alias{get_shooting_list}
\title{get the information}
\usage{
get_shooting_list(path_to_excel, remove_empty = T, ...)
}
\arguments{
\item{path_to_excel}{String. Path to the excel table.}

\item{remove_empty}{Boolean. If \code{TRUE} remove the empty table in the list.}
}
\value{
list of data.table with the edited excel tables.
}
\description{
Read the excel tables and edit them.
}
\examples{
get_shoting_list(path_to_excel = "Desktop/excel/")
}
