% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{get_the_one_list}
\alias{get_the_one_list}
\title{Save excel-tables in one list .}
\usage{
get_the_one_list(path_to_excel, lists_of_excel = LIST_OF_TABLE_FROM_EXCEL,
  remove_empty = T, ...)
}
\arguments{
\item{path_to_excel}{String. Path to the folder of the tables}

\item{remove_empty}{Boolean. If \code{TRUE} remove the empty table in the list.}

\item{list_of_excel}{Character Vector. Names of the excel-tables.}
}
\value{
list.
}
\description{
Save excel-tables in one list .
}
