% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{lese_excel}
\alias{lese_excel}
\title{Read one excel table.}
\usage{
lese_excel(name_excel_table, path_to_excel, format = ".xlsx", ...)
}
\arguments{
\item{name_excel_table}{String. Name of the excel table.}

\item{path_to_excel}{String. Path to the excel table.}

\item{format}{String. Ending of the filename.}
}
\value{
data.table
}
\description{
For given path, read the excel table with name \code{name_excel_table}.
}
