% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_variables.R
\name{load_LIST_OF_TABLE_FROM_EXCEL}
\alias{load_LIST_OF_TABLE_FROM_EXCEL}
\title{save namelist of the excel tables in the package datafile}
\usage{
load_LIST_OF_TABLE_FROM_EXCEL()
}
\value{
Boolean.
}
\description{
save namelist of the excel tables in the package datafile
}
