% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_variables.R
\name{load_NAMES_EXCEL_TABLE}
\alias{load_NAMES_EXCEL_TABLE}
\title{save namelist of the excel tables in the package datafile}
\usage{
load_NAMES_EXCEL_TABLE()
}
\value{
Boolean.
}
\description{
save namelist of the excel tables in the package datafile
}
