% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_variables.R
\name{load_SHOTS_COLUMNS}
\alias{load_SHOTS_COLUMNS}
\title{save the columnnames of shots in the package datafile}
\usage{
load_SHOTS_COLUMNS()
}
\value{
Boolean.
}
\description{
save the columnnames of shots in the package datafile
}
