% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_shots.R
\name{remove_columns}
\alias{remove_columns}
\title{remove columns from Shots}
\usage{
remove_columns(table_shot)
}
\arguments{
\item{table_shot}{data.table
return data.table}
}
\description{
remove columns from Shots
}
