% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_shots.R
\name{remove_rows}
\alias{remove_rows}
\title{remove rows from Shots}
\usage{
remove_rows(table_shot)
}
\arguments{
\item{table_shot}{data.table
return data.table}
}
\description{
remove rows from Shots
}
