% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{clean_shots}
\alias{clean_shots}
\title{Filter only the information}
\usage{
clean_shots(table_shots, ...)
}
\arguments{
\item{table_shots}{data.table. Shots table}
}
\value{
data.table. with separated columns.
}
\description{
Filter only the information
}
