% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daten_einlesen.R
\name{clean_shots}
\alias{clean_shots}
\title{Filter only the information}
\usage{
clean_shots(x, ...)
}
\arguments{
\item{x}{data.table. Shots}
}
\value{
data.table. with separated columns.
}
\description{
Filter only the information
}
