% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daten_einlesen.R
\name{get_the_one_list}
\alias{get_the_one_list}
\title{Save tables in one list .}
\usage{
get_the_one_list(lists_of_excel = LIST_OF_TABLE_FROM_EXCEL,
  path = PATH_TO_EXCEL_TABLE, remove_empty = REMOVE_EMPTY_TABLE, ...)
}
\arguments{
\item{path}{String. Path to the folder of the tables}

\item{remove_empty}{Boolean.}

\item{list_of_excel}{Character Vector. Names of the excel-tables.}
}
\value{
list.
}
\description{
Save tables in one list .
}
