% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daten_einlesen.R
\name{lese_excel}
\alias{lese_excel}
\title{Read excel table.}
\usage{
lese_excel(name, path, format = ".xlsx", ...)
}
\arguments{
\item{name}{String. Name of the excel table.}

\item{path}{String. Path to the excel table.}

\item{format}{String. Ending of the filename.}
}
\value{
data.table
}
\description{
For given path, read the excel table with name x.
}
